+++ demo/mkinit_demo_pkg/subpkg/__init__.py
@@ -1,0 +1,17 @@
+
+import lazy_loader
+
+
+__getattr__, __dir__, __all__ = lazy_loader.attach(
+    __name__,
+    submodules={
+        'nested',
+    },
+    submod_attrs={
+        'nested': [
+            'nested_func',
+        ],
+    },
+)
+
+__all__ = ['nested', 'nested_func']