+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Fixed
 * GH Issue #14: async functions are now handled correctly
+* Issue with `EAGER_IMPORT`
 
 ### Changed
 * Drop support for Python 2.7 and 3.5