+++ CHANGELOG.md
@@ -5,7 +5,7 @@ We are currently working on porting this changelog to the specifications in
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## Version: 1.0.0
+## Version: 1.0.0 - Released 2022-12-03
 
 ### Added
 * Can now specify `"*"` in `__submodules__` to pattern match on available submodules.